move: 파일 위치 controller 안으로 이동
